Moved out static stuff into partial class...keep the DTO clean
